fix scoreboard

+++ Unity/Assets/Script/Contract/PlayerScore.cs
@@ -1,9 +1,9 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-
+using System.Numerics;
 public class PlayerScore {
-    public int value;
+    public long value;
     public int timestamp;
     public string name;
     public int durationMillis;
